Cast defines for EPOLLIN/EPOLLOUT explicitely before negating bits.

+++ src/linux/cio_linux_epoll.c
@@ -101,7 +101,7 @@ enum cio_error cio_linux_eventloop_register_read(const struct cio_eventloop *loo
 
 enum cio_error cio_linux_eventloop_unregister_read(const struct cio_eventloop *loop, struct cio_event_notifier *ev)
 {
-	ev->registered_events &= ~EPOLLIN;
+	ev->registered_events &= ~(uint32_t)EPOLLIN;
 	return epoll_mod(loop, ev, ev->registered_events);
 }
 
@@ -113,7 +113,7 @@ enum cio_error cio_linux_eventloop_register_write(const struct cio_eventloop *lo
 
 enum cio_error cio_linux_eventloop_unregister_write(const struct cio_eventloop *loop, struct cio_event_notifier *ev)
 {
-	ev->registered_events &= ~EPOLLOUT;
+	ev->registered_events &= ~(uint32_t)EPOLLOUT;
 	return epoll_mod(loop, ev, ev->registered_events);
 }
